api/requirements.txt
faiss-cpu>=1.7.4
langchain>=0.3.18
langchain-core>=0.3.34
langchain-openai>=0.0.10
langchain-community>=0.3.17
python-dotenv>=1.0.0
requests==2.31.0
flask>=3.0.0
gunicorn>=21.2.0